Extract per-stop arrivals core into shared functions
The arrivals-and-departures-for-stop handler carried its entire
pipeline inline: the +/-1 day service window scan, the batch route and
trip resolution, the per-row arrival construction, and the reference
assembly. Nothing was callable from anywhere else, so a second endpoint
needing arrivals for many stops would have to duplicate all of it.

Move that pipeline into arrivals_core.go behind arrivalsForStop and
buildArrivalsReferences, with an arrivalsAccumulator gathering the
routes, trips, stops and situations that the arrivals reference. The
accumulator is what makes the code reusable across several stops: a
caller loops over stops sharing one accumulator and gets a single
deduplicated references block.

Behavior is unchanged; the test suite passes without modification.

Two details worth noting for review:

BuildTripStatus now receives the vehicle the caller already looked up
rather than nil. It is not a fix -- BuildTripStatus falls back to the
same GetVehicleForTrip call when handed nil -- so the result is
identical, but it drops one redundant lookup per arrival row.

The extracted stop reference deliberately keeps its inline literal
instead of calling buildStopModel. That helper defaults Code to the
stop ID when stops.code is NULL, where this endpoint emits an empty
string, so adopting it would be a response change rather than a
refactor.
